todo_app/requirements.txt
-i https://pypi.org/simple
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.9.0; python_version >= '3.9'
click==8.2.1; python_version >= '3.10'
fastapi==0.116.1; python_version >= '3.8'
h11==0.16.0; python_version >= '3.8'
httptools==0.6.4; python_full_version >= '3.8.0'
idna==3.10; python_version >= '3.6'
pydantic==2.11.7; python_version >= '3.9'
pydantic-core==2.33.2; python_version >= '3.9'
python-dotenv==1.1.1; python_version >= '3.9'
pyyaml==6.0.2; python_version >= '3.8'
sniffio==1.3.1; python_version >= '3.7'
starlette==0.47.2; python_version >= '3.9'
typing-extensions==4.14.1; python_version >= '3.9'
typing-inspection==0.4.1; python_version >= '3.9'
uvicorn[standard]==0.35.0; python_version >= '3.9'
uvloop==0.21.0; python_full_version >= '3.8.0'
watchfiles==1.1.0; python_version >= '3.9'
websockets==15.0.1; python_version >= '3.9'
-i https://pypi.org/simple
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.9.0; python_version >= '3.9'
click==8.2.1; python_version >= '3.10'
fastapi==0.116.1; python_version >= '3.8'
greenlet==3.2.3; python_version >= '3.9'
h11==0.16.0; python_version >= '3.8'
httptools==0.6.4; python_full_version >= '3.8.0'
idna==3.10; python_version >= '3.6'
pydantic==2.11.7; python_version >= '3.9'
pydantic-core==2.33.2; python_version >= '3.9'
python-dotenv==1.1.1; python_version >= '3.9'
pyyaml==6.0.2; python_version >= '3.8'
sniffio==1.3.1; python_version >= '3.7'
sqlalchemy==2.0.42; python_version >= '3.7'
starlette==0.47.2; python_version >= '3.9'
typing-extensions==4.14.1; python_version >= '3.9'
typing-inspection==0.4.1; python_version >= '3.9'
uvicorn[standard]==0.35.0; python_version >= '3.9'
uvloop==0.21.0; python_full_version >= '3.8.0'
watchfiles==1.1.0; python_version >= '3.9'
websockets==15.0.1; python_version >= '3.9'
-i https://pypi.org/simple
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.9.0; python_version >= '3.9'
bcrypt==4.0.1; python_version >= '3.6'
click==8.2.1; python_version >= '3.10'
fastapi==0.116.1; python_version >= '3.8'
greenlet==3.2.3; python_version >= '3.9'
h11==0.16.0; python_version >= '3.8'
httptools==0.6.4; python_full_version >= '3.8.0'
idna==3.10; python_version >= '3.6'
passlib==1.7.4
pydantic==2.11.7; python_version >= '3.9'
pydantic-core==2.33.2; python_version >= '3.9'
python-dotenv==1.1.1; python_version >= '3.9'
pyyaml==6.0.2; python_version >= '3.8'
sniffio==1.3.1; python_version >= '3.7'
sqlalchemy==2.0.42; python_version >= '3.7'
starlette==0.47.2; python_version >= '3.9'
typing-extensions==4.14.1; python_version >= '3.9'
typing-inspection==0.4.1; python_version >= '3.9'
uvicorn[standard]==0.35.0; python_version >= '3.9'
uvloop==0.21.0; python_full_version >= '3.8.0'
watchfiles==1.1.0; python_version >= '3.9'
websockets==15.0.1; python_version >= '3.9'
-i https://pypi.org/simple
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.9.0; python_version >= '3.9'
bcrypt==4.0.1; python_version >= '3.6'
click==8.2.1; python_version >= '3.10'
fastapi==0.116.1; python_version >= '3.8'
greenlet==3.2.3; python_version >= '3.9'
h11==0.16.0; python_version >= '3.8'
httptools==0.6.4; python_full_version >= '3.8.0'
idna==3.10; python_version >= '3.6'
passlib==1.7.4
pydantic==2.11.7; python_version >= '3.9'
pydantic-core==2.33.2; python_version >= '3.9'
python-dotenv==1.1.1; python_version >= '3.9'
python-multipart==0.0.20; python_version >= '3.8'
pyyaml==6.0.2; python_version >= '3.8'
sniffio==1.3.1; python_version >= '3.7'
sqlalchemy==2.0.42; python_version >= '3.7'
starlette==0.47.2; python_version >= '3.9'
typing-extensions==4.14.1; python_version >= '3.9'
typing-inspection==0.4.1; python_version >= '3.9'
uvicorn[standard]==0.35.0; python_version >= '3.9'
uvloop==0.21.0; python_full_version >= '3.8.0'
watchfiles==1.1.0; python_version >= '3.9'
websockets==15.0.1; python_version >= '3.9'
-i https://pypi.org/simple
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.9.0; python_version >= '3.9'
bcrypt==4.0.1; python_version >= '3.6'
cffi==1.17.1; python_version >= '3.8'
click==8.2.1; python_version >= '3.10'
cryptography==45.0.5; python_version >= '3.7' and python_full_version not in '3.9.0, 3.9.1'
ecdsa==0.19.1; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
fastapi==0.116.1; python_version >= '3.8'
greenlet==3.2.3; python_version >= '3.9'
h11==0.16.0; python_version >= '3.8'
httptools==0.6.4; python_full_version >= '3.8.0'
idna==3.10; python_version >= '3.6'
passlib==1.7.4
pyasn1==0.6.1; python_version >= '3.8'
pycparser==2.22; python_version >= '3.8'
pydantic==2.11.7; python_version >= '3.9'
pydantic-core==2.33.2; python_version >= '3.9'
python-dotenv==1.1.1; python_version >= '3.9'
python-jose[cryptography]==3.5.0; python_version >= '3.9'
python-multipart==0.0.20; python_version >= '3.8'
pyyaml==6.0.2; python_version >= '3.8'
rsa==4.9.1; python_version >= '3.6' and python_version < '4'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sniffio==1.3.1; python_version >= '3.7'
sqlalchemy==2.0.42; python_version >= '3.7'
starlette==0.47.2; python_version >= '3.9'
typing-extensions==4.14.1; python_version >= '3.9'
typing-inspection==0.4.1; python_version >= '3.9'
uvicorn[standard]==0.35.0; python_version >= '3.9'
uvloop==0.21.0; python_full_version >= '3.8.0'
watchfiles==1.1.0; python_version >= '3.9'
websockets==15.0.1; python_version >= '3.9'
